Move ImmArraySeq back into ImmArray
changelog_begin
changelog_end
