+++ daml-lf/data/src/test/scala/com/digitalasset/daml/lf/data/LawlessTraversalsSpec.scala
@@ -21,11 +21,11 @@ class LawlessTraversalsSpec
     }
 
     "preserve class if the implementation bothered to set it up" in {
-      val classySeqs = Seq[Seq[Int]](List(1), Vector(2), ImmArraySeq(3))
+      val classySeqs = Seq[Seq[Int]](List(1), Vector(2), ImmArray.ImmArraySeq(3))
       // we need to use patmat, not == or shouldBe, because patmat is stricter
       inside(classySeqs map (_ traverseEitherStrictly (Right(_)))) {
         case Seq(_, Right(List(_)), _) => fail("lists are not vectors")
-        case Seq(Right(List(1)), Right(Vector(2)), Right(ImmArraySeq(3))) =>
+        case Seq(Right(List(1)), Right(Vector(2)), Right(ImmArray.ImmArraySeq(3))) =>
       }
     }
   }
