+++ daml-lf/transaction-test-lib/src/main/scala/lf/value/test/TypedValueGenerators.scala
@@ -7,7 +7,8 @@ package value
 package test
 
 import scala.language.higherKinds
-import data.{FrontStack, ImmArray, ImmArraySeq, ImmArrayCons, Numeric, Ref, SortedLookupList, Time}
+import data.{FrontStack, ImmArray, ImmArrayCons, Numeric, Ref, SortedLookupList, Time}
+import ImmArray.ImmArraySeq
 import data.DataArbitrary._
 import iface.{
   DefDataType,
