+++ language-support/scala/codegen/src/main/scala/com/digitalasset/codegen/Util.scala
@@ -5,7 +5,8 @@ package com.daml.codegen
 
 import com.daml.codegen.dependencygraph.{OrderedDependencies, TypeDeclOrTemplateWrapper}
 import com.daml.lf.iface.{Type => IType, _}
-import com.daml.lf.data.{ImmArraySeq, Ref}
+import com.daml.lf.data.Ref
+import com.daml.lf.data.ImmArray.ImmArraySeq
 
 import java.io.File
