+++ language-support/scala/codegen/src/main/scala/com/digitalasset/codegen/dependencygraph/DependencyGraph.scala
@@ -7,7 +7,7 @@ import com.daml.lf.iface._
 import com.daml.lf.iface.reader.InterfaceType
 import com.daml.lf.data.Ref.Identifier
 import com.daml.codegen.{Util, lf}
-import com.daml.lf.data.ImmArraySeq
+import com.daml.lf.data.ImmArray.ImmArraySeq
 import lf.DefTemplateWithRecord
 import scalaz.std.list._
 import scalaz.syntax.bifoldable._
