+++ ledger-service/http-json/src/main/scala/com/digitalasset/http/PackageService.scala
@@ -3,7 +3,8 @@
 
 package com.daml.http
 
-import com.daml.lf.data.{ImmArraySeq, Ref}
+import com.daml.lf.data.ImmArray.ImmArraySeq
+import com.daml.lf.data.Ref
 import com.daml.lf.iface
 import com.daml.http.domain.{Choice, TemplateId}
 import com.daml.http.util.IdentifierConverters
