+++ ledger-service/http-json/src/main/scala/com/digitalasset/http/query/ValuePredicate.scala
@@ -5,7 +5,8 @@ package com.daml.http
 package query
 
 import util.IdentifierConverters.lfIdentifier
-import com.daml.lf.data.{ImmArraySeq, Numeric, Ref, Time, Utf8}
+import com.daml.lf.data.{ImmArray, Numeric, Ref, Time, Utf8}
+import ImmArray.ImmArraySeq
 import com.daml.lf.data.ScalazEqual._
 import com.daml.lf.iface
 import com.daml.lf.value.json.JsonVariant
